Move Appelflap modules, change Canoe to Bero
