+++ src/riot/Screens/Sync.riot.html
@@ -47,9 +47,9 @@
 
     <script>
         import { AppDataStatus } from "ts/AppDataStatus";
-        import { AppelflapConnect } from "ts/AppelflapConnect";
-        import { CacheUtilities } from "ts/CacheUtilities";
-        import { CachePublish } from "ts/CachePublish";
+        import { AppelflapConnect } from "ts/Appelflap/AppelflapConnect";
+        import { CacheUtilities } from "ts/Appelflap/CacheUtilities";
+        import { CachePublish } from "ts/Appelflap/CachePublish";
 
         export default {
             state: {
