+++ src/ts/Appelflap/AppelflapConnect.ts
@@ -5,7 +5,7 @@ import {
     TPublications,
     TPublicationTarget,
     TSubscriptions,
-} from "./Types/CacheTypes";
+} from "../Types/CacheTypes";
 
 /* eslint-disable prettier/prettier */
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
@@ -14,7 +14,7 @@ import { AF_CERTCHAIN_LENGTH_HEADER, AF_LOCALHOSTURI, APPELFLAPCOMMANDS, Appelfl
 // The above import statement MUST all appear on the one line for the @ts-ignore to work
 /* eslint-enable prettier/prettier */
 
-import Logger from "./Logger";
+import Logger from "../Logger";
 
 const logger = new Logger("AppelflapConnect");
