+++ src/ts/Appelflap/StartUp.ts
@@ -1,42 +1,42 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
-import { CanoeHost } from "./CanoeHost";
+import { BeroHost } from "./BeroHost";
 import { CertChain } from "./CertChain";
-import Logger from "./Logger";
+import Logger from "../Logger";
 
 const logger = new Logger("StartUp");
 
-/** Initialise the CanoeHost (a wrapper around Appelflap)
+/** Initialise the BeroHost (a wrapper around Appelflap)
  * @param { function } startup - an optional void returning function that is performed after the attempt to start Appelflap.
  */
-export const InitialiseCanoeHost = async (
+export const InitialiseBeroHost = async (
     startup: () => void = () => {}
 ): Promise<string> => {
     try {
         const gt = globalThis as Record<string, any>;
-        gt.canoeHost = new CanoeHost();
-        await gt.canoeHost.StartCanoe(startup);
-        logger.info("CanoeHost initialised");
+        gt.beroHost = new BeroHost();
+        await gt.beroHost.StartBero(startup);
+        logger.info("BeroHost initialised");
         return await Promise.resolve("");
     } catch (e) {
-        logger.info("CanoeHost not initialised, could not achieve 'lock'");
+        logger.info("BeroHost not initialised, could not achieve 'lock'");
         return await Promise.resolve(
-            "Appelflap apparently running, but could not achieve 'lock' for Canoe"
+            "Appelflap apparently running, but could not achieve 'lock' for Bero"
         );
     }
 };
 
 /** Initialise the certificate chain, this should only work if:
- * - Canoe is hosted in Appelflap
+ * - Bero is hosted in Appelflap
  * - The user is logged in
  * - The user has the correct permissions to publish
  */
 export const initialiseCertChain = async (): Promise<void> => {
     const gt = globalThis as Record<string, any>;
-    if (!gt.canoeHost || gt.certChain) {
+    if (!gt.beroHost || gt.certChain) {
         return;
     }
 
-    const afc = gt.canoeHost.appelflapConnect;
+    const afc = gt.beroHost.appelflapConnect;
 
     if (afc && !gt.certChain) {
         try {
@@ -46,7 +46,7 @@ export const initialiseCertChain = async (): Promise<void> => {
                 throw new Error(
                     "Publishing certificate chain could not be initialised. \
                     The certificate endpoint may not be working, \
-                    Canoe is not running within Appelflap (the mobile host), \
+                    Bero is not running within Appelflap (the mobile host), \
                     the user is not logged in \
                     or the user may not have the correct permissions."
                 );
